# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NEON KJ karaoke show management API
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface Show {
  id?: number;
  name?: string;
  startedAt?: string | null;
  endedAt?: string | null;
  isActive?: boolean;
}

export interface Singer {
  id?: number;
  showId?: number;
  name?: string;
  virtualStageTimeMinutes?: number;
  songsSung?: number;
  joinedAt?: string;
}

export interface Song {
  id?: number;
  title?: string;
  artist?: string;
  bpm?: number | null;
  genre?: string | null;
  durationMinutes?: number;
  energyScore?: number;
  filePath?: string | null;
  fileType?: string | null;
}

export interface QueueEntry {
  id?: number;
  showId?: number;
  singerId?: number;
  songId?: number;
  status?: string;
  position?: number;
  addedAt?: string;
}

export interface QueueItem {
  entryId?: number;
  singerId?: number;
  singerName?: string;
  songId?: number;
  songTitle?: string;
  songArtist?: string;
  songDurationMinutes?: number;
  songEnergyScore?: number;
  songBpm?: number | null;
  songGenre?: string | null;
  virtualStageTimeMinutes?: number;
  effectiveStageTimeMinutes?: number;
  position?: number;
  waitTimeMinutes?: number;
  status?: string;
  addedAt?: string;
}

export interface QueueState {
  queue?: QueueItem[];
  nowPlaying?: QueueItem | null;
  showId?: number | null;
  showName?: string | null;
  isPeakHours?: boolean;
  lowEnergyAlert?: boolean;
}

export interface ShowState {
  show?: Show | null;
  queue?: QueueItem[];
  nowPlaying?: QueueItem | null;
  isPeakHours?: boolean;
  lowEnergyAlert?: boolean;
}

export interface QrCodeResponse {
  url?: string;
  qrDataUrl?: string;
}

export type CreateShowBody = {
  name: string;
};

export type EndShow200 = {
  success?: boolean;
};

export type RegisterSingerBody = {
  name: string;
};

export type SearchSongsParams = {
  q?: string;
};

export type UpdateSongBody = {
  bpm?: number;
  energyScore?: number;
};

export type AddToQueueBody = {
  singerId: number;
  songId: number;
};

export type ReorderQueueBody = {
  orderedIds: number[];
};

export type ReorderQueue200 = {
  success?: boolean;
};

export type SkipQueueEntry200 = {
  success?: boolean;
};

export type RemoveFromQueue200 = {
  success?: boolean;
};

export type GetStatus200 = {
  wsClients?: number;
};