# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: NEON KJ karaoke show management API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: shows
  - name: singers
  - name: songs
  - name: queue
  - name: utils

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /shows:
    post:
      operationId: createShow
      tags: [shows]
      summary: Start a new karaoke show
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        "201":
          description: Show created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Show"

  /shows/current:
    get:
      operationId: getCurrentShow
      tags: [shows]
      summary: Get the active show and its queue
      responses:
        "200":
          description: Current show state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShowState"

  /shows/end:
    post:
      operationId: endShow
      tags: [shows]
      summary: End the active show
      responses:
        "200":
          description: Show ended
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean

  /singers:
    post:
      operationId: registerSinger
      tags: [singers]
      summary: Register a singer for the active show
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        "201":
          description: Singer registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Singer"
    get:
      operationId: listSingers
      tags: [singers]
      summary: List all singers in the active show
      responses:
        "200":
          description: Singers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Singer"

  /songs:
    get:
      operationId: searchSongs
      tags: [songs]
      summary: Search the song library
      parameters:
        - in: query
          name: q
          schema:
            type: string
      responses:
        "200":
          description: Song results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Song"

  /songs/{id}:
    get:
      operationId: getSong
      tags: [songs]
      summary: Get a song by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Song
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
    patch:
      operationId: updateSong
      tags: [songs]
      summary: Override BPM or energy score
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bpm:
                  type: integer
                energyScore:
                  type: integer
      responses:
        "200":
          description: Updated song
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"

  /queue:
    get:
      operationId: getQueue
      tags: [queue]
      summary: Get the rotation queue with wait times
      responses:
        "200":
          description: Queue state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueState"
    post:
      operationId: addToQueue
      tags: [queue]
      summary: Add a singer and song to the queue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [singerId, songId]
              properties:
                singerId:
                  type: integer
                songId:
                  type: integer
      responses:
        "201":
          description: Added to queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueEntry"

  /queue/advance:
    post:
      operationId: advanceQueue
      tags: [queue]
      summary: Mark current performer done, bring up next singer
      responses:
        "200":
          description: Queue advanced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueState"

  /queue/reorder:
    patch:
      operationId: reorderQueue
      tags: [queue]
      summary: Manually reorder queue entries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orderedIds]
              properties:
                orderedIds:
                  type: array
                  items:
                    type: integer
      responses:
        "200":
          description: Reordered
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean

  /queue/{id}/skip:
    post:
      operationId: skipQueueEntry
      tags: [queue]
      summary: Skip a singer in the queue
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Skipped
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean

  /queue/{id}:
    delete:
      operationId: removeFromQueue
      tags: [queue]
      summary: Remove a singer from the queue
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean

  /qrcode:
    get:
      operationId: getQrCode
      tags: [utils]
      summary: Get a QR code for the singer join URL
      responses:
        "200":
          description: QR code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCodeResponse"

  /status:
    get:
      operationId: getStatus
      tags: [utils]
      summary: Server status
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema:
                type: object
                properties:
                  wsClients:
                    type: integer

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    Show:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        startedAt:
          type: string
          nullable: true
        endedAt:
          type: string
          nullable: true
        isActive:
          type: boolean

    Singer:
      type: object
      properties:
        id:
          type: integer
        showId:
          type: integer
        name:
          type: string
        virtualStageTimeMinutes:
          type: number
        songsSung:
          type: integer
        joinedAt:
          type: string

    Song:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        artist:
          type: string
        bpm:
          type: integer
          nullable: true
        genre:
          type: string
          nullable: true
        durationMinutes:
          type: number
        energyScore:
          type: integer
        filePath:
          type: string
          nullable: true
        fileType:
          type: string
          nullable: true

    QueueEntry:
      type: object
      properties:
        id:
          type: integer
        showId:
          type: integer
        singerId:
          type: integer
        songId:
          type: integer
        status:
          type: string
        position:
          type: integer
        addedAt:
          type: string

    QueueItem:
      type: object
      properties:
        entryId:
          type: integer
        singerId:
          type: integer
        singerName:
          type: string
        songId:
          type: integer
        songTitle:
          type: string
        songArtist:
          type: string
        songDurationMinutes:
          type: number
        songEnergyScore:
          type: integer
        songBpm:
          type: integer
          nullable: true
        songGenre:
          type: string
          nullable: true
        virtualStageTimeMinutes:
          type: number
        effectiveStageTimeMinutes:
          type: number
        position:
          type: integer
        waitTimeMinutes:
          type: number
        status:
          type: string
        addedAt:
          type: string

    QueueState:
      type: object
      properties:
        queue:
          type: array
          items:
            $ref: "#/components/schemas/QueueItem"
        nowPlaying:
          $ref: "#/components/schemas/QueueItem"
          nullable: true
        showId:
          type: integer
          nullable: true
        showName:
          type: string
          nullable: true
        isPeakHours:
          type: boolean
        lowEnergyAlert:
          type: boolean

    ShowState:
      type: object
      properties:
        show:
          $ref: "#/components/schemas/Show"
          nullable: true
        queue:
          type: array
          items:
            $ref: "#/components/schemas/QueueItem"
        nowPlaying:
          $ref: "#/components/schemas/QueueItem"
          nullable: true
        isPeakHours:
          type: boolean
        lowEnergyAlert:
          type: boolean

    QrCodeResponse:
      type: object
      properties:
        url:
          type: string
        qrDataUrl:
          type: string