# lib/api-zod/src/generated/types/queueItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NEON KJ karaoke show management API
 * OpenAPI spec version: 1.0.0
 */

export interface QueueItem {
  entryId?: number;
  singerId?: number;
  singerName?: string;
  songId?: number;
  songTitle?: string;
  songArtist?: string;
  songDurationMinutes?: number;
  songEnergyScore?: number;
  songBpm?: number | null;
  songGenre?: string | null;
  virtualStageTimeMinutes?: number;
  effectiveStageTimeMinutes?: number;
  position?: number;
  waitTimeMinutes?: number;
  status?: string;
  addedAt?: string;
}