# lib/api-zod/src/generated/types/queueState.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NEON KJ karaoke show management API
 * OpenAPI spec version: 1.0.0
 */
import type { QueueItem } from "./queueItem";

export interface QueueState {
  queue?: QueueItem[];
  nowPlaying?: QueueItem | null;
  showId?: number | null;
  showName?: string | null;
  isPeakHours?: boolean;
  lowEnergyAlert?: boolean;
}