# lib/api-zod/src/generated/types/show.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NEON KJ karaoke show management API
 * OpenAPI spec version: 1.0.0
 */

export interface Show {
  id?: number;
  name?: string;
  startedAt?: string | null;
  endedAt?: string | null;
  isActive?: boolean;
}