# lib/api-zod/src/generated/types/singer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NEON KJ karaoke show management API
 * OpenAPI spec version: 1.0.0
 */

export interface Singer {
  id?: number;
  showId?: number;
  name?: string;
  virtualStageTimeMinutes?: number;
  songsSung?: number;
  joinedAt?: string;
}